backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-dotenv>=1.0.1
openai>=1.51.0
python-docx>=1.1.2
tiktoken>=0.7.0
pydantic>=2.10.0
